+++ packages/dnb-ui-lib/src/components/toggle-button/index.js
@@ -1,0 +1,8 @@
+/**
+ * Component Entry
+ *
+ */
+
+import ToggleButton from './ToggleButton'
+export default ToggleButton
+export * from './ToggleButton'